그룹이 존재한다면 타 그룹 참가 불가

+++ src/components/modals/submodals/WithdrawalModal.js
@@ -5,24 +5,29 @@ import { PROXY } from '../../../contexts/ProxyContext';
 
 const WithdrawalModal=(props)=>{
     function removeInfo(){
-        axios({
-            method:'post',
-            url:`${PROXY}/member/unlink`,
-            data:{
-                accessToken: localStorage.getItem('access_token'),
-            },
-            headers:{
-                'ContentType':'application/json'
-            },
-        })
-        .then(() => {
-            localStorage.clear();
-            window.location.replace('/');
-        })
-        .catch(() => {
-            alert("탈퇴에 실패하였습니다.");
-            window.location.replace('/main');
-        })
+        if(localStorage.getItem("mygroupid")){
+            alert("현재 그룹이 존재합니다.\n그룹을 퇴장하시고 진행해주세요.");
+        }
+        else{
+            axios({
+                method:'post',
+                url:`${PROXY}/member/unlink`,
+                data:{
+                    accessToken: localStorage.getItem('access_token'),
+                },
+                headers:{
+                    'ContentType':'application/json'
+                },
+            })
+            .then(() => {
+                localStorage.clear();
+                window.location.replace('/');
+            })
+            .catch(() => {
+                alert("탈퇴에 실패하였습니다.");
+                window.location.replace('/main');
+            })
+        }
     }
 
     return(
@@ -44,7 +49,7 @@ const WithdrawalModal=(props)=>{
 
             <Modal.Footer className='WDFooter' style={{backgroundColor: "#FFFCEE"}}>
                 <Button variant="dark" className='WDButtonLeft' onClick={props.onHide}>취소</Button>
-                <Button variant="lgiht" className='WDButtonRight' onClick={removeInfo}>회원 탈퇴</Button>
+                <Button variant="lgiht" className='WDButtonRight' onClick={()=>{removeInfo(); props.onHide()}}>회원 탈퇴</Button>
             </Modal.Footer>
         </Modal>
     );
